# src/main/java/it/bitify/libreria/model/entity/Student.java
package it.bitify.libreria.model.entity;

import jakarta.persistence.*;

import java.util.List;
import java.util.Set;

@Entity
@Table(name = "student")
public class Student {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "id")
    private Long id;

    @Column(name = "name", nullable = false, length = 30)
    private String name;

    @Column(name = "surname", nullable = false, length = 30)
    private String surname;

    @Column(name = "class", nullable = false, length = 2)
    private String schoolClass;

    @Column(name = "email", nullable = false, length = 30, unique = true)
    private String email;

    @OneToMany(mappedBy = "student", cascade = {CascadeType.PERSIST, CascadeType.MERGE, CascadeType.DETACH, CascadeType.REFRESH})
    private List<Loan> loans;

    @OneToOne(cascade = CascadeType.ALL)
    @JoinColumn(name="idCard")
    private Card card;

    @ManyToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE, CascadeType.DETACH, CascadeType.REFRESH})
    @JoinTable(name = "ENROLLMENT", joinColumns = @JoinColumn(name = "idStudent"),
            inverseJoinColumns = @JoinColumn(name = "idCourse"))
    private Set<Course> courses;

    public Long getStudentId() {
        return id;
    }

    public void setStudentId(Long id) {
        this.id = id;
    }

    public String getsurname() {
        return surname;
    }

    public void setsurname(String surname) {
        this.surname = surname;
    }

    public String getClasse() {
        return schoolClass;
    }

    public void setClasse(String schoolClass) {
        this.schoolClass = schoolClass;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public Card getCard() {
        return card;
    }

    public void setCard(Card card) {
        this.card = card;
    }

    public String getname() {
        return name;
    }

    public void setname(String name) {
        this.name = name;
    }
}