# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Libreria API definition
  version: v1
servers:
- url: http://localhost:8080
  description: base server url
...
paths:
...
  /api/v1/loan:
    get:
      tags:
      - Loan
      - get-all
      summary: Retrieve all loans
      description: Retrieve all Loan objects. The response is a list of Loan objects
      operationId: getAll_1
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Retrieved all Loans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageLoan"
    put:
      tags:
      - Loan
      - put
      summary: Modify Loan
      description: Modify a Loan object by specifying its id
      operationId: editLoan
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Loan"
        required: true
      responses:
        "400":
          description: Id type mismatch
        "404":
          description: Loan not found
        "500":
          description: Data integrity violation
    post:
      tags:
      - post
      - Loan
      summary: Add Loan
      description: Add a Loan object to the database
      operationId: addLoan
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Loan"
        required: true
      responses:
        "200":
          description: Loan added to the database
        "500":
          description: Data integrity violation
  /api/v1/loan/{id}:
    get:
      tags:
        - Loan
        - get
      summary: Retrieve Loan by id
      description: Retrieve a Loan object by specifying its id. The response is a
        Loan object
      operationId: getLoanById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Loan found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "400":
          description: Id type mismatch
        "404":
          description: Loan not found
    delete:
      tags:
        - Loan
        - delete
      summary: Delete Loan by id
      description: Delete a Loan object by specifying its id
      operationId: deleteLoan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Loan deleted
        "400":
          description: Id type mismatch
        "404":
          description: Loan not found
        "500":
          description: Data integrity violation
  /api/v1/loan/ongoing:
    get:
      tags:
        - Loan
        - get-all
      summary: Retrieve all ongoing loans
      description: Retrieve all ongoing Loan objects. The response is a list of Loan objects with null endDate
      operationId: getOngoing
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Retrieved all ongoing Loans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageLoan"
...
  /api/v1/course:
    get:
      tags:
      - get-all
      - Course
      summary: Retrieve all courses
      description: Retrieve all Course objects. The response is a list of Course objects
      operationId: getAll_3
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Retrieved all Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageCourse"
    put:
      tags:
      - Course
      - put
      summary: Modify Course
      description: Modify a Course object by specifying its id
      operationId: editCourse
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Course"
        required: true
      responses:
        "200":
          description: Course modified
        "400":
          description: Id type mismatch
        "404":
          description: Course not found
        "500":
          description: Data integrity violation
    post:
      tags:
      - post
      - Course
      summary: Add Course
      description: Add a Course object to the database
      operationId: addCourse
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Course"
        required: true
      responses:
        "200":
          description: Course added to the database
        "500":
          description: Data integrity violation
  /api/v1/course/{id}:
    get:
      tags:
        - get
        - Course
      summary: Retrieve Course by id
      description: Retrieve a Course object by specifying its id. The response is
        a Course object
      operationId: getCourseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Course found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "400":
          description: Id type mismatch
        "404":
          description: Course not found
    delete:
      tags:
        - Course
        - delete
      summary: Delete Course by id
      description: Delete a Course object by specifying its id
      operationId: deleteCourse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Course deleted
        "400":
          description: Id type mismatch
        "404":
          description: Course not found
        "500":
          description: Data integrity violation
  /api/v1/category:
    get:
      tags:
      - Category
      - get-all
      summary: Retrieve all categories
      description: Retrieve all Category objects. The response is a list of Category objects
        objects
      operationId: getAllCategories
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Retrieved all Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageCategory"
    put:
      tags:
      - Category
      - put
      summary: Edit Category
      description: Edit a Category object by specifying its id
      operationId: editCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
        required: true
      responses:
        "200":
          description: Category modified
        "400":
          description: Id type mismatch
        "404":
          description: Category not found
        "500":
          description: Data integrity violation
    post:
      tags:
      - Category
      - post
      summary: Add Category
      description: Add a Category object to the database
      operationId: addCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
        required: true
      responses:
        "200":
          description: Category added to the database
        "500":
          description: Data integrity violation
  /api/v1/category/{id}:
    get:
      tags:
      - Category
      - get
      summary: Retrieve Category by id
      description: Retrieve a Category object by specifying its id. The response is a Category object
      operationId: getCategoryById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Category found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Id type mismatch
        "404":
          description: Category not found
    delete:
      tags:
      - Category
      - delete
      summary: Delete Category by id
      description: Delete a Category object by specifying its id
      operationId: removeCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Category deleted
        "400":
          description: Id type mismatch
        "404":
          description: Category not found
        "500":
          description: Data integrity violation
  /api/v1/category/book-count:
    get:
      tags:
        - Category
        - get-all
      summary: Retrieve all categories with book count
      description: Retrieve all Category objects. The response is a list of Category names with the associated book count
        objects
      operationId: getAllCategoriesWithBookCount
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Retrieved all Categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageCategory"
  /api/v1/category/only-class:
    get:
      tags:
        - Category
        - get-all
      summary: Retrieve all categories with book lent to students belonging to only one class
      description: Retrieve all categories with book lent to students belonging to only one class
      operationId: getCategoriesOnlyClass
      parameters:
        - name: targetClass
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Retrieved all Categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageCategory"
        "400":
          description: id type mismatch
components:
  schemas:
    Card:
      type: object
      properties:
        id:
          type: integer
          format: int64
        releaseDate:
          type: string
          format: date
    Student:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        class:
          type: string
        email:
          type: string
        Card:
          $ref: "#/components/schemas/Card"
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
        year:
          type: integer
          format: int32
        category:
          $ref: "#/components/schemas/Category"
        title:
          type: string
        author:
          type: string
    Loan:
      type: object
      properties:
        id:
          type: integer
          format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        student:
          $ref: "#/components/schemas/Student"
        Book:
          $ref: "#/components/schemas/Book"
    Course:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        students:
          type: array
          items:
            $ref: "#/components/schemas/Student"
          uniqueItems: true
...
    PageLoan:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Loan'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
    PageCourse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Course'
        totalElements:
          type: integer
        totalPages:
          type: integerdate
        size:
          type: integer
        number:
          type: integer
    PageCategory:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
...
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 5
    AscParam:
      name: ascending
      in: query
      required: false
      schema:
        type: boolean
        default: true
    SortByParam:
      name: sortBy
      in: query
      required: false
      schema:
        type: string
        default: id